SOFTWARE_TESIS/ARSAC/Backend/ARSAC/requirements.txt
Django==4.2.1
gunicorn == 20.0.4 
psycopg2-binary==2.8.5
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.3
PyJWT==2.7.0
pytz==2023.3